EMO-7580 sonarcube fixes

+++ blob/src/main/java/com/bazaarvoice/emodb/blob/db/s3/S3RateLimiter.java
@@ -140,7 +140,7 @@ public class S3RateLimiter {
         private void waitForRetry(Throwable t, Method method) throws Throwable {
             // Backoff for a random amount of time between 1 and 5 seconds
             try {
-                Thread.sleep(1000 + (int) (Math.random() * 4000));
+                Thread.sleep(1000 + (long) (Math.random() * 4000));
             } catch (InterruptedException e) {
                 // On interrupt don't keep retrying, just throw the original exception
                 LOGGER.warn("S3 operation interrupted while retrying rate limited request");
